backend/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django_extensions
psycopg2
python-dotenv
django-nanoid-field
drf-spectacular
django-filter
google-auth
requests
pillow
